+++ docs/MAUI/ReleaseNotes.md
@@ -1,4 +1,4 @@
-## PortableShop MAUI: ReleaseNotes 2.x
+## PortableShop - MAUI: ReleaseNotes 2.x
 
 #### Summary
 Working MAUI application. Still limited to Android release.
@@ -6,12 +6,13 @@ Working MAUI application. Still limited to Android release.
 #### What's new
 * Layout improvements.
 * Set & restore window dimensions on Windows.
+* Localization now working on both platforms.
 
 #### Known issues
 * Not yet released for Windows.
 * No splash screen on Windows.
-* Localization not working on Windows.
 * Display shopping cart incorrect on Windows, until resizing.
 * Validation of text filter not properly working.
 * Refresh button invisible on Windows.
 * Various icons and titles not working, depending on platform and theme setting.
+* Restored filter values not visible on Windows.